Add files via upload
FRAUD DETECTION DATASET
To find the accuracy of an dataset. First we are supposed to load and then Split the data.. Splitting the data helps in training and testing the dataset.
Then adding the input features and train Regression which is to be performed  in the data.
Predicting and final step to print the needed output ie. Accuracy.